+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\codeworkspace\\cloud-single-page-website\\src\\index.js":"1","C:\\codeworkspace\\cloud-single-page-website\\src\\App.js":"2","C:\\codeworkspace\\cloud-single-page-website\\src\\reportWebVitals.js":"3","C:\\codeworkspace\\cloud-single-page-website\\src\\stringSplit.js":"4"},{"size":556,"mtime":1610455029967,"results":"5","hashOfConfig":"6"},{"size":4378,"mtime":1610474021366,"results":"7","hashOfConfig":"6"},{"size":362,"mtime":499162500000,"results":"8","hashOfConfig":"6"},{"size":1089,"mtime":1610473074810,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1nnm43j",{"filePath":"12","messages":"13","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\codeworkspace\\cloud-single-page-website\\src\\index.js",[],"C:\\codeworkspace\\cloud-single-page-website\\src\\App.js",["19"],"C:\\codeworkspace\\cloud-single-page-website\\src\\reportWebVitals.js",[],["20","21"],"C:\\codeworkspace\\cloud-single-page-website\\src\\stringSplit.js",[],{"ruleId":"22","severity":1,"message":"23","line":30,"column":6,"nodeType":"24","endLine":30,"endColumn":8,"suggestions":"25"},{"ruleId":"26","replacedBy":"27"},{"ruleId":"28","replacedBy":"29"},"react-hooks/exhaustive-deps","React Hook useCallback has a missing dependency: 'textItself'. Either include it or remove the dependency array.","ArrayExpression",["30"],"no-native-reassign",["31"],"no-negated-in-lhs",["32"],{"desc":"33","fix":"34"},"no-global-assign","no-unsafe-negation","Update the dependencies array to be: [textItself]",{"range":"35","text":"36"},[1129,1131],"[textItself]"]
+[{"C:\\codeworkspace\\cloud-single-page-website\\src\\index.js":"1","C:\\codeworkspace\\cloud-single-page-website\\src\\App.js":"2","C:\\codeworkspace\\cloud-single-page-website\\src\\reportWebVitals.js":"3","C:\\codeworkspace\\cloud-single-page-website\\src\\stringSplit.js":"4","C:\\codeworkspace\\cloud-single-page-website\\src\\Components\\Title.js":"5","C:\\codeworkspace\\cloud-single-page-website\\src\\Components\\DropZone.js":"6","C:\\codeworkspace\\cloud-single-page-website\\src\\Components\\TableOfWords.js":"7","C:\\codeworkspace\\cloud-single-page-website\\src\\Components\\TextContent.js":"8"},{"size":556,"mtime":1610455029967,"results":"9","hashOfConfig":"10"},{"size":520,"mtime":1610482657397,"results":"11","hashOfConfig":"10"},{"size":362,"mtime":499162500000,"results":"12","hashOfConfig":"10"},{"size":1089,"mtime":1610473074810,"results":"13","hashOfConfig":"10"},{"size":405,"mtime":1610475671507,"results":"14","hashOfConfig":"10"},{"size":2227,"mtime":1610482689364,"results":"15","hashOfConfig":"10"},{"size":1159,"mtime":1610481222273,"results":"16","hashOfConfig":"10"},{"size":1043,"mtime":1610482640645,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1nnm43j",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\codeworkspace\\cloud-single-page-website\\src\\index.js",[],"C:\\codeworkspace\\cloud-single-page-website\\src\\App.js",[],"C:\\codeworkspace\\cloud-single-page-website\\src\\reportWebVitals.js",[],["35","36"],"C:\\codeworkspace\\cloud-single-page-website\\src\\stringSplit.js",[],"C:\\codeworkspace\\cloud-single-page-website\\src\\Components\\Title.js",[],"C:\\codeworkspace\\cloud-single-page-website\\src\\Components\\DropZone.js",["37"],"C:\\codeworkspace\\cloud-single-page-website\\src\\Components\\TableOfWords.js",[],"C:\\codeworkspace\\cloud-single-page-website\\src\\Components\\TextContent.js",[],{"ruleId":"38","replacedBy":"39"},{"ruleId":"40","replacedBy":"41"},{"ruleId":"42","severity":1,"message":"43","line":32,"column":6,"nodeType":"44","endLine":32,"endColumn":8,"suggestions":"45"},"no-native-reassign",["46"],"no-negated-in-lhs",["47"],"react-hooks/exhaustive-deps","React Hook useCallback has a missing dependency: 'textItself'. Either include it or remove the dependency array.","ArrayExpression",["48"],"no-global-assign","no-unsafe-negation",{"desc":"49","fix":"50"},"Update the dependencies array to be: [textItself]",{"range":"51","text":"52"},[1174,1176],"[textItself]"]
